bugfixing some implicit typecasting

+++ libs/ml/examples/w2v_google_code/main.cpp
@@ -788,7 +788,7 @@ void *TrainModelThread(void *id)
       // linearly in number of words seen, but thresholded below
       // at one ten-thousandth of initial learning rate)
       // (each word in the data is to be seen `iter` times)
-      alpha = starting_alpha * (1 - word_count_actual / static_cast<float>(static_cast<float>(iter * train_words) + 1));
+      alpha = static_cast<float>(starting_alpha) * (static_cast<float>(1) - static_cast<float>(word_count_actual) / (static_cast<float>(iter * train_words) + 1));
       if (alpha < starting_alpha * 0.0001)
         alpha = static_cast<float>(starting_alpha * 0.0001);
     }
@@ -945,11 +945,11 @@ void *TrainModelThread(void *id)
             for (c = 0; c < layer1_size; c++)
               f += neu1[c] * syn1neg[c + l2];
             if (f > MAX_EXP)
-              g = static_cast<float>(label - 1) * alpha;
+              g = (static_cast<float>(label) - static_cast<float>(1)) * alpha;
             else if (f < -MAX_EXP)
-              g = static_cast<float>(label - 0) * alpha;
+              g = (static_cast<float>(label) - static_cast<float>(0)) * alpha;
             else
-              g = (label - static_cast<float>(expTable[static_cast<int>((f + MAX_EXP) * (EXP_TABLE_SIZE / MAX_EXP / 2))])) * alpha;
+              g = (static_cast<float>(label) - static_cast<float>(expTable[static_cast<int>((f + MAX_EXP) * (EXP_TABLE_SIZE / MAX_EXP / 2))])) * alpha;
             for (c = 0; c < layer1_size; c++)
               neu1e[c] += g * syn1neg[c + l2];
             for (c = 0; c < layer1_size; c++)
